Add a call attribute for resource list allocations
Even though this is rather an ownership-transfer-attribute than a call
attribute, it is not specific enough to be tagged to a specific
argument.

While functions with that new attribute will typically register the
resource returned as their primary return value with a resource list, it
is also possible that the allocated resource is not returned at all but
rather used internally.

Another reason is that we ran out of available bits for ownership
transfer attributes, where there was still a bit available for call
attributes.

Now all bits are used up; so let's try to be happy about it!