New branch with UUID l7eabiq3jl9qjq36kjpgzpmum
This branch contains documents establishing a set of conventions for
naming C functions and their parameters.

The purpose of those naming conventions is to annotate functions and
parameters with attributes for which the C language does not provide any
of declaring them.

If the author of some set of source files chooses to follow that set of
conventions, a lot of interface documentation for individual functions
and parameters becomes unnecessary, because the name of the function of
parameter already conveys that information.

The documents of this project are kept in a sub-directory
"./design/features/". The idea is that other similar projects describing
other conventions also put their files there.

This will then be merged into other projects which keep their own design
documents in "./design", and contain the merged conventions in the
"features" sub-directory of that directory.

This set of conventions has initially been developed for the SXC, DCM
and R4G projects, but has been singled out of there because it has been
considered generally useful, and is not tied to any of the original
projects in particular.

The next commit after this one will merge a re-based and filtered
version history of the aforementioned ancestor projects, populating this
new project.