Migrate to asciidoc
There is little to do as the documentation is already almost in
"asciifile" source syntax.

Add a POSIX-compliant Makefile for building HTML documentation.

Add more default assumptions what values to consider "not valid" in
absence of explicit documentation for "optional" ownership-transfer
parameters.

Rebase the project to include a branch UUID at its first commit.

Prepare a version bump to version 1.0 of the specification.